ensure stubs get the right name
Sometimes Merger decides to rename an internal function that will have
its stub put in the stub module. LoadPartDefinition() can add a
declaration under the new name to the stub module before we actually
create the stub function. So when we create the stub function, we have
to replace anything already using the name.