Client: use delays to work with the new memodb-server
The new server always responds immediately to requests, even when no
result is available yet. That means there's no need to create lots of
HTTP connections waiting for long requests. We can just use one
connection per thread.

When evaluateAsync() is called, it immediately sends the first request
to the server, to make sure the job is queued. After that, it doesn't
send any more requests until the result is actually needed.