ClientEvaluator: don't create so many threads
The old design tried to create new threads whenever existing threads
were waiting for a result, in order to make sure there were always
threads doing useful work. But each thread needed its own connection,
and it was running out of sockets.

The new design keeps a constant number of threads, and threads waiting
for a result will also check if there's any work they can do while
waiting. The downside is that we can end up with an arbitrary number of
recursive calls, causing us to run out of stack space, but hopefully
that will never happen in practice.